fix(tui): stabilize tab status detail updates
